Layout basic home structure.
