+++ src/pages/home/home.js
@@ -1,6 +1,46 @@
 import React from 'react';
+// internal
+import style from './style.module.css';
 
 
 export default function Home() {
-  return <h3>Home</h3>
+  return (
+    <React.Fragment>
+      <section id='' className={`${style.section} content-layout`}>
+        <div>
+          <h3>Some section title</h3>
+        </div>
+        <div className='light-bg'>
+        </div>
+      </section>
+      <section id='' className={`${style.section} content-layout`}>
+        <div>
+          <h3>Some section title</h3>
+        </div>
+        <div className='light-bg'>
+        </div>
+      </section>
+      <section id='' className={`${style.section} content-layout`}>
+        <div>
+          <h3>Some section title</h3>
+        </div>
+        <div className='light-bg'>
+        </div>
+      </section>
+      <section id='' className={`${style.section} content-layout`}>
+        <div>
+          <h3>Some section title</h3>
+        </div>
+        <div className='light-bg'>
+        </div>
+      </section>
+      <section id='' className={`${style.section} content-layout`}>
+        <div>
+          <h3>Some section title</h3>
+        </div>
+        <div className='light-bg'>
+        </div>
+      </section>
+    </React.Fragment>
+  );
 }
