+++ src/pages/home/style.module.css
@@ -1,0 +1,12 @@
+section > div{
+  display: block;
+  height: calc(3 * var(--large));
+  grid-column-start: 3;
+  grid-column-end: 4;
+  grid-row: 1;
+}
+section > div:first-of-type{
+  grid-column-start: 2;
+  grid-column-end: 3;
+  grid-row: 1;
+}
